Update Raku_Steering_Council_Code.md
The original document proposed here was copied with minor alterations, but contains excess words, unclear statements, poor grammar, and refers to concepts that are not defined. The original document was created - apparently - to deal with problems in another software community. Whilst copying is good, there is no need to copy obvious mistakes. 
There are changes I have made 
* to remove unnecessary words and clarify grammar
* to clarify aspects of elections based on experience
* to change emotive language, such as expulsion, to more widely accepted electoral processes, such as recall votes. The processes described in the original proposal
are confusing and would allow someone in the core group to demand constant recalls. Specifying when a recall can take place for a member of the Council or for the whole Council means a recall only happens for cause, not out of pique. When there is a generally perceived need for the whole council to be recalled, then something is truly wrong, so a new Council will be needed in any case, and the people demanding the change would probably want to be elected in their place. 
* to add the Raku philosophy, basically be minimally editing Damian Conway's excellent response to issue #3. 
* to add explicitly a Code of Conduct, based on the original text, responses in the issues. I have also explicitly included the Netiquette rules of Virginia Shea as they are widely quoted for internet interactions and cover almost all virtual interactions.
* I have added a power to the Council to declare some people Core development team. For example, I cannot imagine that Larry Wall, Damian Conway, Pierre Michaud (I think other names should be added that I haven't done) should ever be removed from the Core development team, even if they not longer make comits. The existing proposal would force such a change.