codesections nomination
Based on issue #20, I believe that my commit-bit to the raku/doc
repository allows me to nominate myself to serve on the Raku Steering
Council.  If I am incorrect about that, please feel free to close this
PR.

This PR also updates the voting form to include list my name and
github handle.