Clarify Raku Steering Council Code
The current Raku Steering Council Code combines details of the initial
proposal for adopting the code with actual implementation details.
This commit revises the Code to remove the portions that are not
relevant post-adoption (but preserves the initial proposal for future
reference).  It also provides definitions for a few terms that were
used but not defined in the prior version.  One such term is "Core
Team"; as part of that definition, this commit adds a separate
document listing the Core Team's current membership.

The changes in this commit are all intended to be non-substantive,
with one minor exception: the previous version of the Code said that
Core Team members would be deemed to be inactive unless they either
(1) voted in the initial election for the RSC or (2) had made a commit
in the two years before the RSC was first adopted.  This commit
changes (2) to be voted in the ratification of these changes to the
RSC Code.  This makes it much easier to determine who is a currently
active member of the Core Team and doesn't really matter, since both
versions of the Code state that an inactive Core Team member can
become active "at will".